// src/main/vnpay/vnpay.dto.ts
export class VnpayDto {
  vnp_SecureHash: string;
  vnp_SecureHashType: string;
  vnp_ResponseCode: string;
  vnp_Amount: string;
  vnp_BankCode: string;
  vnp_BankTranNo: string;
  vnp_CardType: string;
  vnp_OrderInfo: string;
  vnp_PayDate: string;
  vnp_TmnCode: string;
  vnp_TransactionNo: string;
  vnp_TransactionStatus: string;
  vnp_TxnRef: string;
}